Various fixes:
 * fix compatibility with newer silver-platter
 * split out github module
 * stop signing pypi uploads - see
   https://blog.pypi.org/posts/2023-05-23-removing-pgp/
